Switch to IPStack's IP-to-country resolution
IP Vigilant is now dysfunctional, the domain is gone and shows a registrar's placeholder.

This, in turn, puts a load on the IP Stack's API usage, so the caching must be enforced (from minutes to days and months per IP, assuming the IP's country does not change often).

Also, avoid querying if there is no IP to query about.